smartline-api/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
rapidfuzz==3.5.2